Update device function tests
---
