--- tests/gpu/device_code2.cpp
@@ -1,4 +1,0 @@
-#include <cstdio>
-#include "gpu_common.h"
-
-__device__ void device_function2(int a) { printf("device_func2 %d\n", a); }
