+++ tests/gpu/device_func.cpp
